ISSUE #274 * Internal refactor `minos.common.protocol.avro`.
