+++ minos/common/protocol/avro/__init__.py
@@ -5,12 +5,12 @@ This file is part of minos framework.
 
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
-from .protocols import (
-    MinosAvroDatabaseProtocol,
-    MinosAvroMessageProtocol,
+from .base import (
     MinosAvroProtocol,
 )
-from .schemas import (
-    DATABASE_AVRO_SCHEMA,
-    MESSAGE_AVRO_SCHEMA,
+from .databases import (
+    MinosAvroDatabaseProtocol,
+)
+from .messages import (
+    MinosAvroMessageProtocol,
 )
