+++ minos/common/protocol/avro/messages.py
@@ -1,0 +1,119 @@
+"""
+Copyright (C) 2021 Clariteia SL
+
+This file is part of minos framework.
+
+Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
+"""
+
+from typing import (
+    Any,
+)
+
+from ...logs import (
+    log,
+)
+from .base import (
+    MinosAvroProtocol,
+)
+
+
+class MinosAvroMessageProtocol(MinosAvroProtocol):
+    """TODO"""
+
+    @classmethod
+    def encode(cls, headers: dict, body: Any = None, *args, **kwargs) -> bytes:
+        """
+        encoder in avro
+        all the headers are converted in fields with double underscore name
+        the body is a set fields coming from the data type.
+        """
+
+        # prepare the headers
+        final_data = dict()
+        final_data["headers"] = headers
+        if body:
+            final_data["body"] = body
+
+        return super().encode(final_data, _AVRO_SCHEMA)
+
+    @classmethod
+    def decode(cls, data: bytes, *args, **kwargs) -> dict:
+        """TODO
+
+        :param data: TODO
+        :param args: TODO
+        :param kwargs: TODO
+        :return: TODO
+        """
+        data_return = {"headers": dict()}
+        for schema_dict in super().decode(data, flatten=False):
+            log.debug("Avro: get the request/response in dict format")
+            data_return["headers"] = schema_dict["headers"]
+            # check wich type is body
+            if "body" in schema_dict:
+                if isinstance(schema_dict["body"], dict):
+                    data_return["body"] = {}
+                elif isinstance(schema_dict["body"], list):
+                    data_return["body"] = []
+                data_return["body"] = schema_dict["body"]
+            return data_return
+
+
+_AVRO_SCHEMA = {
+    "type": "record",
+    "namespace": "org.minos.protocol",
+    "name": "message",
+    "fields": [
+        {"name": "headers", "type": {"type": "map", "values": ["string", "int", "bytes", "long", "float", "boolean"]}},
+        {
+            "name": "body",
+            "type": [
+                "null",
+                "string",
+                "int",
+                "bytes",
+                "boolean",
+                "float",
+                "long",
+                {"type": "array", "items": ["int", "string", "bytes", "long", "boolean", "float"]},
+                {
+                    "type": "map",
+                    "values": [
+                        "null",
+                        "int",
+                        "string",
+                        "bytes",
+                        "boolean",
+                        "float",
+                        "long",
+                        {
+                            "type": "array",
+                            "items": [
+                                "string",
+                                "int",
+                                "bytes",
+                                "long",
+                                "boolean",
+                                "float",
+                                {"type": "map", "values": ["string", "int", "bytes", "long", "boolean", "float"]},
+                            ],
+                        },
+                        {
+                            "type": "map",
+                            "values": [
+                                "string",
+                                "int",
+                                "bytes",
+                                "long",
+                                "boolean",
+                                "float",
+                                {"type": "map", "values": ["string", "int", "bytes", "long", "boolean", "float"]},
+                            ],
+                        },
+                    ],
+                },
+            ],
+        },
+    ],
+}
